Added admin page and some fix + new func

+++ back/controllers/orderController.js
@@ -1,5 +1,7 @@
 const {Orders_Content, Basket_Content,Baskets, Orders, Products, Users} = require('../models/models');
 const ApiError = require("../error/ApiError");
+const e = require("express");
+const sequelize = require("express");
 
 class basketController {
     async createOrder(req, res, next) {
@@ -13,33 +15,53 @@ class basketController {
             }
             const order_content = await Basket_Content.findAll({
                 where: {basketIdBasket: basket.id_basket},
-                include: {model: Products, attributes: ['price']},
+                include: {model: Products, attributes: ['price', 'quantity', 'id']},
             })
             if (order_content.length === 0) {
                 return next(ApiError.badRequest("Ваша корзина пуста!"))
             }
-            const order = await Orders.create({
-                total_price: basket.total_price,
-                userId: id,
-                date: Date.now(),
-            })
-            const orderItem = []
-            for (const item of order_content) {
-                const orderItem = await Orders_Content.create({
-                    orderIdOrder: order.id_order,
-                    productId: item.productId,
-                    quantity: item.quantity,
-                })
-            }
-            await Basket_Content.destroy({
-                where: { basketIdBasket: basket.id_basket }
-            });
 
-            await Baskets.update(
-                { total_price: 0 },
-                { where: { id_basket: basket.id_basket } }
-            );
-            return res.json({message: "Успешно создан заказ!"})
+            let isPossible = true
+            for (const item of order_content) {
+                if (item.quantity > item.product.quantity) {
+                    isPossible = false
+                }
+            }
+
+            if (isPossible) {
+                const order = await Orders.create({
+                    total_price: basket.total_price,
+                    userId: id,
+                    date: Date.now(),
+                })
+
+                const orderItem = []
+                for (const item of order_content) {
+                    await Orders_Content.create({
+                        orderIdOrder: order.id_order,
+                        productId: item.productId,
+                        quantity: item.quantity,
+                    })
+                    console.log(item.toJSON())
+                    let newQuantity = item.product.quantity - item.quantity
+                    await Products.update(
+                        {quantity: (newQuantity)},
+                        {where: {
+                            id: item.productId
+                        }}
+                    )
+                }
+                await Basket_Content.destroy({
+                    where: { basketIdBasket: basket.id_basket }
+                });
+
+                await Baskets.update(
+                    { total_price: 0 },
+                    { where: { id_basket: basket.id_basket } }
+                );
+                return res.json({message: "Успешно создан заказ!"})
+            }
+            return next(ApiError.internal("Товара нет на складе"));
         } catch (e) {
             return next(ApiError.internal(e.message));
         }
